# backend/openapi.yaml
openapi: 3.1.0
info:
  title: RMG ERP System
  version: 1.0.0
  description: |
    Southern Apparels ERP System API
    Multi-Database Architecture with 5 separate PostgreSQL databases
    - Clients DB: Buyers, Suppliers, Contacts, Shipping, Banking
    - Samples DB: Styles, Variants, Samples, TNA, Plans, Operations, Materials, Colors, Sizes
    - Users DB: User management & Authentication
    - Orders DB: Order management
    - Merchandiser DB: Future merchandiser features
servers:
  - url: http://localhost:8000
    description: Development server
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            example:
              email: "john.doe@rmg.com"
              username: "john_doe"
              full_name: "John Doe"
              department: "Merchandising"
              designation: "Senior Merchandiser"
              department_access: ["samples", "orders"]
              password: "SecurePass123"
      responses:
        '201':
          description: User created
          content:
            application/json:
              example:
                id: 1
                email: "john.doe@rmg.com"
                username: "john_doe"
                full_name: "John Doe"
                department: "Merchandising"
                designation: "Senior Merchandiser"
                department_access: ["samples", "orders"]
                is_active: true
                is_superuser: false
                created_at: "2024-01-15T10:30:00Z"
        '422': { description: Validation error }
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Login and get access token
      requestBody:
        required: true
        content:
          application/json:
            example:
              username: "admin"
              password: "admin123"
      responses:
        '200':
          description: Token returned
          content:
            application/json:
              example:
                access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJhZG1pbiIsImV4cCI6MTcwNTMyMDAwMH0.abc123"
                token_type: "bearer"
        '401': { description: Invalid credentials }
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Get current user info
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: User info
          content:
            application/json:
              example:
                id: 1
                email: "admin@rmg.com"
                username: "admin"
                full_name: "System Admin"
                department: "IT"
                designation: "Administrator"
                department_access: ["all"]
                is_active: true
                is_superuser: true
                created_at: "2024-01-01T00:00:00Z"
        '401': { description: Unauthorized }

  # ============ BUYERS ============
...
  /api/v1/orders/:
    get:
      tags: [orders]
      summary: List all orders
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              example:
                - id: 1
                  order_no: "ORD-2024-001"
                  style_name: "Classic Polo"
                  buyer_id: 1
                  style_id: 1
                  season: "SS24"
                  order_category: "Main Order"
                  sales_contract: "SC-2024-001"
                  scl_po: "PO-12345"
                  fob: 12.50
                  note: "First shipment"
                  product_category: "Knit"
                  style_description: "Classic fit polo"
                  gauge: "18GG"
                  is_set: false
                  order_quantity: 5000
                  unit_price: 12.50
                  total_value: 62500.00
                  order_date: "2024-01-01"
                  delivery_date: "2024-03-15"
                  shipment_date: "2024-03-10"
                  order_status: "Confirmed"
                  created_at: "2024-01-01T00:00:00Z"
    post:
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            example:
              order_no: "ORD-2024-002"
              style_name: "Cable Sweater"
              buyer_id: 1
              style_id: 2
              season: "FW24"
              order_category: "Main Order"
              sales_contract: "SC-2024-002"
              scl_po: "PO-12346"
              fob: 25.00
              product_category: "Knit"
              gauge: "7GG"
              is_set: false
              order_quantity: 3000
              unit_price: 25.00
              total_value: 75000.00
              order_date: "2024-01-15"
              delivery_date: "2024-04-30"
              order_status: "Pending"
      responses:
        '201':
          description: Order created
  /api/v1/orders/{order_id}:
    get:
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Order details
        '404': { description: Not found }
    put:
      tags: [orders]
      summary: Update order
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            example:
              order_status: "In Production"
              note: "Production started"
      responses:
        '200':
          description: Order updated
    delete:
      tags: [orders]
      summary: Delete order
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '204': { description: Deleted }

  # ============ MATERIALS ============
...
  /api/v1/users/:
    get:
      tags: [users]
      summary: List all users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              example:
                - id: 1
                  email: "admin@rmg.com"
                  username: "admin"
                  full_name: "System Admin"
                  department: "IT"
                  designation: "Administrator"
                  department_access: ["all"]
                  is_active: true
                  is_superuser: true
                  created_at: "2024-01-01T00:00:00Z"
    post:
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            example:
              email: "merchandiser@rmg.com"
              username: "merchandiser1"
              full_name: "John Merchandiser"
              department: "Merchandising"
              designation: "Senior Merchandiser"
              department_access: ["samples", "orders", "buyers"]
              password: "SecurePass123"
      responses:
        '201':
          description: User created
  /api/v1/users/{user_id}:
    get:
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: User details
        '404': { description: Not found }
    put:
      tags: [users]
      summary: Update user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            example:
              full_name: "Senior Admin"
              designation: "IT Manager"
              is_active: true
      responses:
        '200':
          description: User updated
    delete:
      tags: [users]
      summary: Delete user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '204': { description: Deleted }

  # ============ OPERATIONS ============
...